A new way that makes "cabal test" work.
So you're going to have trouble testing stuff that's in a module that
you want to call "Main", because that's always going to interfere with
the module "Main" that's for your testrunner. So I renamed it and fixed up
some other things. (cabal.sandbox.config shouldn't be checked in, probably)